backend/pyproject.toml
[tool.poetry]
name = "mail-sync-backend"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
fastapi = "^0.109.0"
dependency-injector = "^4.41.0"
httpx = "^0.24.1"
overrides = "^7.3.1"
structlog = "^23.1.0"
asgi-correlation-id = "^4.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pylint = "^2.17.5"
autoflake = "^2.2.0"
pytest = "^7.4.0"
mypy = "^1.8.0"
pre-commit = "^3.3.3"
pytest-cov = "^4.1.0"
editorconfig-checker = "^2.7.2"
debugpy = "^1.6.7"
pytest-httpx = "^0.22.0"
flake8-print = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
float_to_top = true

[tool.black]
line-length = 120
include = '\.pyi?$'

[tool.pylint]
ignore = '.venv'
fail-under = 7
max-line-length = 120
disable = [
  "missing-function-docstring",
  "missing-module-docstring",
  "missing-class-docstring",
]
extension-pkg-allow-list = [
  "pydantic",
  "dependency_injector.providers",
  "dependency_injector.containers",
]
